backend/requirements.txt
# FastAPI & Server
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# Database
motor==3.7.1
pymongo==4.12.1

# Authentication & Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# Validation & Settings
pydantic[email-validator]==2.11.3
pydantic-settings==2.9.1
pydantic[email]

# HTTP Client (for TMDB proxy)
httpx==0.28.1

# Email (password reset)
aiosmtplib==3.0.2

# Utils
python-dotenv==1.1.0

# Logging
colorama
